gpu: Share code between gpu-av and shader printf

gpu: Move vma init into shared code


gpu: Add PreCallRecordCreateDevice to shared code


gpu: Move PostCallRecordCreateDevice to shared code


gpu: Move Create*Pipeline to shared code


gpu: Move message generation to shared code


gpu: Move ProcessInstrumentationBuffer into shared code


gpu: Move SubmitBarrier into shared code


gpu: Add error message back

In PostCallRecordCreateDevice
