+++ build-android/cmake/layerlib/CMakeLists.txt
@@ -70,6 +70,7 @@ add_library(VkLayer_khronos_validation SHARED
         ${SRC_DIR}/layers/buffer_validation.cpp
         ${SRC_DIR}/layers/shader_validation.cpp
         ${SRC_DIR}/layers/gpu_validation.cpp
+        ${SRC_DIR}/layers/gpu_shared.cpp
         ${SRC_DIR}/layers/shader_debug_printf.cpp
         ${SRC_DIR}/layers/best_practices_utils.cpp
         ${COMMON_DIR}/include/layer_chassis_dispatch.cpp
