+++ layers/gpu_validation.cpp
@@ -18,22 +18,10 @@
  * Author: Tony Barbour <tony@lunarg.com>
  */
 
-// This define indicates to build the VMA routines themselves
-#define VMA_IMPLEMENTATION
-// This define indicates that we will supply Vulkan function pointers at initialization
-#define VMA_STATIC_VULKAN_FUNCTIONS 0
 #include "gpu_validation.h"
-#include "shader_validation.h"
-#include "spirv-tools/libspirv.h"
 #include "spirv-tools/optimizer.hpp"
 #include "spirv-tools/instrument.hpp"
-#include <SPIRV/spirv.hpp>
-#include <algorithm>
-#include <regex>
 #include "layer_chassis_dispatch.h"
-
-// This is the number of bindings in the debug descriptor set.
-static const uint32_t kNumBindingsInSet = 3;
 
 static const VkShaderStageFlags kShaderStageAllRayTracing =
     VK_SHADER_STAGE_ANY_HIT_BIT_NV | VK_SHADER_STAGE_CALLABLE_BIT_NV | VK_SHADER_STAGE_CLOSEST_HIT_BIT_NV |
@@ -188,191 +176,6 @@ static const uint32_t kComputeShaderSpirv[] = {
     0x000200f8, 0x0000000d, 0x0004003d, 0x00000006, 0x0000006b, 0x00000008, 0x00050080, 0x00000006, 0x0000006c, 0x0000006b,
     0x00000024, 0x0003003e, 0x00000008, 0x0000006c, 0x000200f9, 0x0000000a, 0x000200f8, 0x0000000c, 0x000100fd, 0x00010038};
 
-// Implementation for Descriptor Set Manager class
-GpuAssistedDescriptorSetManager::GpuAssistedDescriptorSetManager(GpuAssisted *dev_data) { dev_data_ = dev_data; }
-
-GpuAssistedDescriptorSetManager::~GpuAssistedDescriptorSetManager() {
-    for (auto &pool : desc_pool_map_) {
-        DispatchDestroyDescriptorPool(dev_data_->device, pool.first, NULL);
-    }
-    desc_pool_map_.clear();
-}
-
-VkResult GpuAssistedDescriptorSetManager::GetDescriptorSet(VkDescriptorPool *desc_pool, VkDescriptorSet *desc_set) {
-    std::vector<VkDescriptorSet> desc_sets;
-    VkResult result = GetDescriptorSets(1, desc_pool, &desc_sets);
-    if (result == VK_SUCCESS) {
-        *desc_set = desc_sets[0];
-    }
-    return result;
-}
-
-VkResult GpuAssistedDescriptorSetManager::GetDescriptorSets(uint32_t count, VkDescriptorPool *pool,
-                                                            std::vector<VkDescriptorSet> *desc_sets) {
-    const uint32_t default_pool_size = kItemsPerChunk;
-    VkResult result = VK_SUCCESS;
-    VkDescriptorPool pool_to_use = VK_NULL_HANDLE;
-
-    if (0 == count) {
-        return result;
-    }
-    desc_sets->clear();
-    desc_sets->resize(count);
-
-    for (auto &pool : desc_pool_map_) {
-        if (pool.second.used + count < pool.second.size) {
-            pool_to_use = pool.first;
-            break;
-        }
-    }
-    if (VK_NULL_HANDLE == pool_to_use) {
-        uint32_t pool_count = default_pool_size;
-        if (count > default_pool_size) {
-            pool_count = count;
-        }
-        const VkDescriptorPoolSize size_counts = {
-            VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
-            pool_count * kNumBindingsInSet,
-        };
-        VkDescriptorPoolCreateInfo desc_pool_info = {};
-        desc_pool_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-        desc_pool_info.pNext = NULL;
-        desc_pool_info.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
-        desc_pool_info.maxSets = pool_count;
-        desc_pool_info.poolSizeCount = 1;
-        desc_pool_info.pPoolSizes = &size_counts;
-        result = DispatchCreateDescriptorPool(dev_data_->device, &desc_pool_info, NULL, &pool_to_use);
-        assert(result == VK_SUCCESS);
-        if (result != VK_SUCCESS) {
-            return result;
-        }
-        desc_pool_map_[pool_to_use].size = desc_pool_info.maxSets;
-        desc_pool_map_[pool_to_use].used = 0;
-    }
-    std::vector<VkDescriptorSetLayout> desc_layouts(count, dev_data_->debug_desc_layout);
-
-    VkDescriptorSetAllocateInfo alloc_info = {VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO, NULL, pool_to_use, count,
-                                              desc_layouts.data()};
-
-    result = DispatchAllocateDescriptorSets(dev_data_->device, &alloc_info, desc_sets->data());
-    assert(result == VK_SUCCESS);
-    if (result != VK_SUCCESS) {
-        return result;
-    }
-    *pool = pool_to_use;
-    desc_pool_map_[pool_to_use].used += count;
-    return result;
-}
-
-void GpuAssistedDescriptorSetManager::PutBackDescriptorSet(VkDescriptorPool desc_pool, VkDescriptorSet desc_set) {
-    auto iter = desc_pool_map_.find(desc_pool);
-    if (iter != desc_pool_map_.end()) {
-        VkResult result = DispatchFreeDescriptorSets(dev_data_->device, desc_pool, 1, &desc_set);
-        assert(result == VK_SUCCESS);
-        if (result != VK_SUCCESS) {
-            return;
-        }
-        desc_pool_map_[desc_pool].used--;
-        if (0 == desc_pool_map_[desc_pool].used) {
-            DispatchDestroyDescriptorPool(dev_data_->device, desc_pool, NULL);
-            desc_pool_map_.erase(desc_pool);
-        }
-    }
-    return;
-}
-
-// Trampolines to make VMA call Dispatch for Vulkan calls
-static VKAPI_ATTR void VKAPI_CALL gpuVkGetPhysicalDeviceProperties(VkPhysicalDevice physicalDevice,
-                                                                   VkPhysicalDeviceProperties *pProperties) {
-    DispatchGetPhysicalDeviceProperties(physicalDevice, pProperties);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkGetPhysicalDeviceMemoryProperties(VkPhysicalDevice physicalDevice,
-                                                                         VkPhysicalDeviceMemoryProperties *pMemoryProperties) {
-    DispatchGetPhysicalDeviceMemoryProperties(physicalDevice, pMemoryProperties);
-}
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkAllocateMemory(VkDevice device, const VkMemoryAllocateInfo *pAllocateInfo,
-                                                          const VkAllocationCallbacks *pAllocator, VkDeviceMemory *pMemory) {
-    return DispatchAllocateMemory(device, pAllocateInfo, pAllocator, pMemory);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkFreeMemory(VkDevice device, VkDeviceMemory memory, const VkAllocationCallbacks *pAllocator) {
-    DispatchFreeMemory(device, memory, pAllocator);
-}
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkMapMemory(VkDevice device, VkDeviceMemory memory, VkDeviceSize offset, VkDeviceSize size,
-                                                     VkMemoryMapFlags flags, void **ppData) {
-    return DispatchMapMemory(device, memory, offset, size, flags, ppData);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkUnmapMemory(VkDevice device, VkDeviceMemory memory) { DispatchUnmapMemory(device, memory); }
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkFlushMappedMemoryRanges(VkDevice device, uint32_t memoryRangeCount,
-                                                                   const VkMappedMemoryRange *pMemoryRanges) {
-    return DispatchFlushMappedMemoryRanges(device, memoryRangeCount, pMemoryRanges);
-}
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkInvalidateMappedMemoryRanges(VkDevice device, uint32_t memoryRangeCount,
-                                                                        const VkMappedMemoryRange *pMemoryRanges) {
-    return DispatchInvalidateMappedMemoryRanges(device, memoryRangeCount, pMemoryRanges);
-}
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkBindBufferMemory(VkDevice device, VkBuffer buffer, VkDeviceMemory memory,
-                                                            VkDeviceSize memoryOffset) {
-    return DispatchBindBufferMemory(device, buffer, memory, memoryOffset);
-}
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkBindImageMemory(VkDevice device, VkImage image, VkDeviceMemory memory,
-                                                           VkDeviceSize memoryOffset) {
-    return DispatchBindImageMemory(device, image, memory, memoryOffset);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkGetBufferMemoryRequirements(VkDevice device, VkBuffer buffer,
-                                                                   VkMemoryRequirements *pMemoryRequirements) {
-    DispatchGetBufferMemoryRequirements(device, buffer, pMemoryRequirements);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkGetImageMemoryRequirements(VkDevice device, VkImage image,
-                                                                  VkMemoryRequirements *pMemoryRequirements) {
-    DispatchGetImageMemoryRequirements(device, image, pMemoryRequirements);
-}
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkCreateBuffer(VkDevice device, const VkBufferCreateInfo *pCreateInfo,
-                                                        const VkAllocationCallbacks *pAllocator, VkBuffer *pBuffer) {
-    return DispatchCreateBuffer(device, pCreateInfo, pAllocator, pBuffer);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkDestroyBuffer(VkDevice device, VkBuffer buffer, const VkAllocationCallbacks *pAllocator) {
-    return DispatchDestroyBuffer(device, buffer, pAllocator);
-}
-static VKAPI_ATTR VkResult VKAPI_CALL gpuVkCreateImage(VkDevice device, const VkImageCreateInfo *pCreateInfo,
-                                                       const VkAllocationCallbacks *pAllocator, VkImage *pImage) {
-    return DispatchCreateImage(device, pCreateInfo, pAllocator, pImage);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkDestroyImage(VkDevice device, VkImage image, const VkAllocationCallbacks *pAllocator) {
-    DispatchDestroyImage(device, image, pAllocator);
-}
-static VKAPI_ATTR void VKAPI_CALL gpuVkCmdCopyBuffer(VkCommandBuffer commandBuffer, VkBuffer srcBuffer, VkBuffer dstBuffer,
-                                                     uint32_t regionCount, const VkBufferCopy *pRegions) {
-    DispatchCmdCopyBuffer(commandBuffer, srcBuffer, dstBuffer, regionCount, pRegions);
-}
-
-VkResult GpuAssisted::InitializeVma(VkPhysicalDevice physical_device, VkDevice device, VmaAllocator *pAllocator) {
-    VmaVulkanFunctions functions;
-    VmaAllocatorCreateInfo allocatorInfo = {};
-    allocatorInfo.device = device;
-    allocatorInfo.physicalDevice = physical_device;
-
-    functions.vkGetPhysicalDeviceProperties = (PFN_vkGetPhysicalDeviceProperties)gpuVkGetPhysicalDeviceProperties;
-    functions.vkGetPhysicalDeviceMemoryProperties = (PFN_vkGetPhysicalDeviceMemoryProperties)gpuVkGetPhysicalDeviceMemoryProperties;
-    functions.vkAllocateMemory = (PFN_vkAllocateMemory)gpuVkAllocateMemory;
-    functions.vkFreeMemory = (PFN_vkFreeMemory)gpuVkFreeMemory;
-    functions.vkMapMemory = (PFN_vkMapMemory)gpuVkMapMemory;
-    functions.vkUnmapMemory = (PFN_vkUnmapMemory)gpuVkUnmapMemory;
-    functions.vkFlushMappedMemoryRanges = (PFN_vkFlushMappedMemoryRanges)gpuVkFlushMappedMemoryRanges;
-    functions.vkInvalidateMappedMemoryRanges = (PFN_vkInvalidateMappedMemoryRanges)gpuVkInvalidateMappedMemoryRanges;
-    functions.vkBindBufferMemory = (PFN_vkBindBufferMemory)gpuVkBindBufferMemory;
-    functions.vkBindImageMemory = (PFN_vkBindImageMemory)gpuVkBindImageMemory;
-    functions.vkGetBufferMemoryRequirements = (PFN_vkGetBufferMemoryRequirements)gpuVkGetBufferMemoryRequirements;
-    functions.vkGetImageMemoryRequirements = (PFN_vkGetImageMemoryRequirements)gpuVkGetImageMemoryRequirements;
-    functions.vkCreateBuffer = (PFN_vkCreateBuffer)gpuVkCreateBuffer;
-    functions.vkDestroyBuffer = (PFN_vkDestroyBuffer)gpuVkDestroyBuffer;
-    functions.vkCreateImage = (PFN_vkCreateImage)gpuVkCreateImage;
-    functions.vkDestroyImage = (PFN_vkDestroyImage)gpuVkDestroyImage;
-    functions.vkCmdCopyBuffer = (PFN_vkCmdCopyBuffer)gpuVkCmdCopyBuffer;
-    allocatorInfo.pVulkanFunctions = &functions;
-
-    return vmaCreateAllocator(&allocatorInfo, pAllocator);
-}
-
 // Convenience function for reporting problems with setting up GPU Validation.
 template <typename T>
 void GpuAssisted::ReportSetupProblem(T object, const char *const specific_message) const {
@@ -394,33 +197,12 @@ void GpuAssisted::PreCallRecordCreateBuffer(VkDevice device, const VkBufferCreat
 void GpuAssisted::PreCallRecordCreateDevice(VkPhysicalDevice gpu, const VkDeviceCreateInfo *create_info,
                                             const VkAllocationCallbacks *pAllocator, VkDevice *pDevice,
                                             safe_VkDeviceCreateInfo *modified_create_info) {
-    VkPhysicalDeviceFeatures supported_features;
     DispatchGetPhysicalDeviceFeatures(gpu, &supported_features);
-    if (supported_features.fragmentStoresAndAtomics || supported_features.vertexPipelineStoresAndAtomics ||
-        supported_features.shaderInt64) {
-        VkPhysicalDeviceFeatures *features = nullptr;
-        if (modified_create_info->pEnabledFeatures) {
-            // If pEnabledFeatures, VkPhysicalDeviceFeatures2 in pNext chain is not allowed
-            features = const_cast<VkPhysicalDeviceFeatures *>(modified_create_info->pEnabledFeatures);
-        } else {
-            VkPhysicalDeviceFeatures2 *features2 = nullptr;
-            features2 =
-                const_cast<VkPhysicalDeviceFeatures2 *>(lvl_find_in_chain<VkPhysicalDeviceFeatures2>(modified_create_info->pNext));
-            if (features2) features = &features2->features;
-        }
-        if (features) {
-            features->fragmentStoresAndAtomics = supported_features.fragmentStoresAndAtomics;
-            features->vertexPipelineStoresAndAtomics = supported_features.vertexPipelineStoresAndAtomics;
-            features->shaderInt64 = supported_features.shaderInt64;
-        } else {
-            VkPhysicalDeviceFeatures new_features = {};
-            new_features.fragmentStoresAndAtomics = supported_features.fragmentStoresAndAtomics;
-            new_features.vertexPipelineStoresAndAtomics = supported_features.vertexPipelineStoresAndAtomics;
-            new_features.shaderInt64 = supported_features.shaderInt64;
-            delete modified_create_info->pEnabledFeatures;
-            modified_create_info->pEnabledFeatures = new VkPhysicalDeviceFeatures(new_features);
-        }
-    }
+    VkPhysicalDeviceFeatures features = {};
+    features.vertexPipelineStoresAndAtomics = true;
+    features.fragmentStoresAndAtomics = true;
+    features.shaderInt64 = true;
+    SharedPreCallRecordCreateDevice(gpu, modified_create_info, supported_features, features);
 }
 // Perform initializations that can be done at Create Device time.
 void GpuAssisted::PostCallRecordCreateDevice(VkPhysicalDevice physicalDevice, const VkDeviceCreateInfo *pCreateInfo,
@@ -453,88 +235,19 @@ void GpuAssisted::PostCallRecordCreateDevice(VkPhysicalDevice physicalDevice, co
                    "shaderInt64 feature is not available.  No buffer device address checking will be attempted");
     }
     device_gpu_assisted->shaderInt64 = device_gpu_assisted->enabled_features.core.shaderInt64;
-
-    // If api version 1.1 or later, SetDeviceLoaderData will be in the loader
-    auto chain_info = get_chain_info(pCreateInfo, VK_LOADER_DATA_CALLBACK);
-    assert(chain_info->u.pfnSetDeviceLoaderData);
-    device_gpu_assisted->vkSetDeviceLoaderData = chain_info->u.pfnSetDeviceLoaderData;
-
-    // Some devices have extremely high limits here, so set a reasonable max because we have to pad
-    // the pipeline layout with dummy descriptor set layouts.
-    device_gpu_assisted->adjusted_max_desc_sets = device_gpu_assisted->phys_dev_props.limits.maxBoundDescriptorSets;
-    device_gpu_assisted->adjusted_max_desc_sets = std::min(33U, device_gpu_assisted->adjusted_max_desc_sets);
-
-    // We can't do anything if there is only one.
-    // Device probably not a legit Vulkan device, since there should be at least 4. Protect ourselves.
-    if (device_gpu_assisted->adjusted_max_desc_sets == 1) {
-        ReportSetupProblem(device, "Device can bind only a single descriptor set.  GPU-Assisted Validation disabled.");
-        device_gpu_assisted->aborted = true;
-        return;
+    device_gpu_assisted->physicalDevice = physicalDevice;
+    device_gpu_assisted->device = *pDevice;
+    device_gpu_assisted->output_buffer_size = sizeof(uint32_t) * (spvtools::kInst2MaxOutCnt + 1);
+    std::vector<VkDescriptorSetLayoutBinding> bindings;
+    VkDescriptorSetLayoutBinding binding = {0, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1,
+                                            VK_SHADER_STAGE_ALL_GRAPHICS | VK_SHADER_STAGE_COMPUTE_BIT | kShaderStageAllRayTracing,
+                                            NULL};
+    bindings.push_back(binding);
+    for (auto i = 1; i < 3; i++) {
+        binding.binding = i;
+        bindings.push_back(binding);
     }
-    device_gpu_assisted->desc_set_bind_index = device_gpu_assisted->adjusted_max_desc_sets - 1;
-    LogInfo(device, "UNASSIGNED-GPU-Assisted Validation. ", "Shaders using descriptor set at index %d. ",
-            device_gpu_assisted->desc_set_bind_index);
-
-    device_gpu_assisted->output_buffer_size = sizeof(uint32_t) * (spvtools::kInstMaxOutCnt + 1);
-    VkResult result1 = InitializeVma(physicalDevice, *pDevice, &device_gpu_assisted->vmaAllocator);
-    assert(result1 == VK_SUCCESS);
-    std::unique_ptr<GpuAssistedDescriptorSetManager> desc_set_manager(new GpuAssistedDescriptorSetManager(device_gpu_assisted));
-
-    // The descriptor indexing checks require only the first "output" binding.
-    const VkDescriptorSetLayoutBinding debug_desc_layout_bindings[kNumBindingsInSet] = {
-        {
-            0,  // output
-            VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
-            1,
-            VK_SHADER_STAGE_ALL_GRAPHICS | VK_SHADER_STAGE_COMPUTE_BIT | kShaderStageAllRayTracing,
-            NULL,
-        },
-        {
-            1,  // input
-            VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
-            1,
-            VK_SHADER_STAGE_ALL_GRAPHICS | VK_SHADER_STAGE_COMPUTE_BIT | kShaderStageAllRayTracing,
-            NULL,
-        },
-        {
-            2,  // input
-            VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
-            1,
-            VK_SHADER_STAGE_ALL_GRAPHICS | VK_SHADER_STAGE_COMPUTE_BIT | kShaderStageAllRayTracing,
-            NULL,
-        },
-    };
-    const VkDescriptorSetLayoutCreateInfo debug_desc_layout_info = {VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO, NULL, 0,
-                                                                    kNumBindingsInSet, debug_desc_layout_bindings};
-
-    const VkDescriptorSetLayoutCreateInfo dummy_desc_layout_info = {VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO, NULL, 0, 0,
-                                                                    NULL};
-
-    result1 = DispatchCreateDescriptorSetLayout(*pDevice, &debug_desc_layout_info, NULL, &device_gpu_assisted->debug_desc_layout);
-
-    // This is a layout used to "pad" a pipeline layout to fill in any gaps to the selected bind index.
-    VkResult result2 =
-        DispatchCreateDescriptorSetLayout(*pDevice, &dummy_desc_layout_info, NULL, &device_gpu_assisted->dummy_desc_layout);
-    assert((result1 == VK_SUCCESS) && (result2 == VK_SUCCESS));
-    if ((result1 != VK_SUCCESS) || (result2 != VK_SUCCESS)) {
-        ReportSetupProblem(*pDevice, "Unable to create descriptor set layout.  GPU-Assisted Validation disabled.");
-        if (result == VK_SUCCESS) {
-            DispatchDestroyDescriptorSetLayout(*pDevice, device_gpu_assisted->debug_desc_layout, NULL);
-        }
-        if (result2 == VK_SUCCESS) {
-            DispatchDestroyDescriptorSetLayout(*pDevice, device_gpu_assisted->dummy_desc_layout, NULL);
-        }
-        device_gpu_assisted->debug_desc_layout = VK_NULL_HANDLE;
-        device_gpu_assisted->dummy_desc_layout = VK_NULL_HANDLE;
-        device_gpu_assisted->aborted = true;
-        return;
-    }
-    device_gpu_assisted->desc_set_manager = std::move(desc_set_manager);
-
-    // Register callback to be called at any ResetCommandBuffer time
-    device_gpu_assisted->SetCommandBufferResetCallback(
-        [device_gpu_assisted](VkCommandBuffer command_buffer) -> void { device_gpu_assisted->ResetCommandBuffer(command_buffer); });
-
+    SharedPostCallRecordCreateDevice(pCreateInfo, bindings, device_gpu_assisted, device_gpu_assisted->phys_dev_props);
     CreateAccelerationStructureBuildValidationState(device_gpu_assisted);
 }
 
@@ -562,35 +275,13 @@ void GpuAssisted::PreCallRecordDestroyBuffer(VkDevice device, VkBuffer buffer, c
     BUFFER_STATE *buffer_state = GetBufferState(buffer);
     if (buffer_state) buffer_map.erase(buffer_state->deviceAddress);
 }
+
 // Clean up device-related resources
 void GpuAssisted::PreCallRecordDestroyDevice(VkDevice device, const VkAllocationCallbacks *pAllocator) {
-    for (auto &queue_barrier_command_info_kv : queue_barrier_command_infos) {
-        GpuAssistedQueueBarrierCommandInfo &queue_barrier_command_info = queue_barrier_command_info_kv.second;
-
-        DispatchFreeCommandBuffers(device, queue_barrier_command_info.barrier_command_pool, 1,
-                                   &queue_barrier_command_info.barrier_command_buffer);
-        queue_barrier_command_info.barrier_command_buffer = VK_NULL_HANDLE;
-
-        DispatchDestroyCommandPool(device, queue_barrier_command_info.barrier_command_pool, NULL);
-        queue_barrier_command_info.barrier_command_pool = VK_NULL_HANDLE;
-    }
-    queue_barrier_command_infos.clear();
-    if (debug_desc_layout) {
-        DispatchDestroyDescriptorSetLayout(device, debug_desc_layout, NULL);
-        debug_desc_layout = VK_NULL_HANDLE;
-    }
-    if (dummy_desc_layout) {
-        DispatchDestroyDescriptorSetLayout(device, dummy_desc_layout, NULL);
-        dummy_desc_layout = VK_NULL_HANDLE;
-    }
-    desc_set_manager.reset();
-
     DestroyAccelerationStructureBuildValidationState();
-
-    if (vmaAllocator) {
-        vmaDestroyAllocator(vmaAllocator);
-    }
+    SharedPreCallRecordDestroyDevice(this);
 }
+
 void GpuAssisted::CreateAccelerationStructureBuildValidationState(GpuAssisted *device_gpuav) {
     if (device_gpuav->aborted) {
         return;
@@ -1138,8 +829,8 @@ void GpuAssisted::PreCallRecordCmdBuildAccelerationStructureNV(VkCommandBuffer c
     static constexpr const VkDeviceSize kInstanceSize = 64;
     const VkDeviceSize instance_buffer_size = kInstanceSize * pInfo->instanceCount;
 
-    result =
-        desc_set_manager->GetDescriptorSet(&as_validation_buffer_info.descriptor_pool, &as_validation_buffer_info.descriptor_set);
+    result = desc_set_manager->GetDescriptorSet(&as_validation_buffer_info.descriptor_pool, debug_desc_layout,
+                                                &as_validation_buffer_info.descriptor_set);
     if (result != VK_SUCCESS) {
         ReportSetupProblem(device, "Unable to get descriptor set for acceleration structure build.");
         aborted = true;
@@ -1269,19 +960,7 @@ void GpuAssisted::PreCallRecordCreatePipelineLayout(VkDevice device, const VkPip
              << "Instrumented shaders are replaced with non-instrumented shaders.";
         ReportSetupProblem(device, strm.str().c_str());
     } else {
-        // Modify the pipeline layout by:
-        // 1. Copying the caller's descriptor set desc_layouts
-        // 2. Fill in dummy descriptor layouts up to the max binding
-        // 3. Fill in with the debug descriptor layout at the max binding slot
-        cpl_state->new_layouts.reserve(adjusted_max_desc_sets);
-        cpl_state->new_layouts.insert(cpl_state->new_layouts.end(), &pCreateInfo->pSetLayouts[0],
-                                      &pCreateInfo->pSetLayouts[pCreateInfo->setLayoutCount]);
-        for (uint32_t i = pCreateInfo->setLayoutCount; i < adjusted_max_desc_sets - 1; ++i) {
-            cpl_state->new_layouts.push_back(dummy_desc_layout);
-        }
-        cpl_state->new_layouts.push_back(debug_desc_layout);
-        cpl_state->modified_create_info.pSetLayouts = cpl_state->new_layouts.data();
-        cpl_state->modified_create_info.setLayoutCount = adjusted_max_desc_sets;
+        SharedPreCallRecordCreatePipelineLayout(cpl_state, this, pCreateInfo);
     }
 }
 
@@ -1301,7 +980,7 @@ void GpuAssisted::ResetCommandBuffer(VkCommandBuffer commandBuffer) {
     if (aborted) {
         return;
     }
-    auto gpuav_buffer_list = GetGpuAssistedBufferInfo(commandBuffer);
+    auto gpuav_buffer_list = GetBufferInfo(commandBuffer);
     for (auto buffer_info : gpuav_buffer_list) {
         vmaDestroyBuffer(vmaAllocator, buffer_info.output_mem_block.buffer, buffer_info.output_mem_block.allocation);
         if (buffer_info.di_input_mem_block.buffer) {
@@ -1377,8 +1056,8 @@ void GpuAssisted::PreCallRecordCreateGraphicsPipelines(VkDevice device, VkPipeli
                                                        void *cgpl_state_data) {
     std::vector<safe_VkGraphicsPipelineCreateInfo> new_pipeline_create_infos;
     create_graphics_pipeline_api_state *cgpl_state = reinterpret_cast<create_graphics_pipeline_api_state *>(cgpl_state_data);
-    PreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, cgpl_state->pipe_state, &new_pipeline_create_infos,
-                                   VK_PIPELINE_BIND_POINT_GRAPHICS);
+    SharedPreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, cgpl_state->pipe_state,
+                                         &new_pipeline_create_infos, VK_PIPELINE_BIND_POINT_GRAPHICS, this);
     cgpl_state->gpu_create_infos = new_pipeline_create_infos;
     cgpl_state->pCreateInfos = reinterpret_cast<VkGraphicsPipelineCreateInfo *>(cgpl_state->gpu_create_infos.data());
 }
@@ -1389,8 +1068,8 @@ void GpuAssisted::PreCallRecordCreateComputePipelines(VkDevice device, VkPipelin
                                                       void *ccpl_state_data) {
     std::vector<safe_VkComputePipelineCreateInfo> new_pipeline_create_infos;
     auto *ccpl_state = reinterpret_cast<create_compute_pipeline_api_state *>(ccpl_state_data);
-    PreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, ccpl_state->pipe_state, &new_pipeline_create_infos,
-                                   VK_PIPELINE_BIND_POINT_COMPUTE);
+    SharedPreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, ccpl_state->pipe_state,
+                                         &new_pipeline_create_infos, VK_PIPELINE_BIND_POINT_COMPUTE, this);
     ccpl_state->gpu_create_infos = new_pipeline_create_infos;
     ccpl_state->pCreateInfos = reinterpret_cast<VkComputePipelineCreateInfo *>(ccpl_state->gpu_create_infos.data());
 }
@@ -1401,8 +1080,8 @@ void GpuAssisted::PreCallRecordCreateRayTracingPipelinesNV(VkDevice device, VkPi
                                                            void *crtpl_state_data) {
     std::vector<safe_VkRayTracingPipelineCreateInfoCommon> new_pipeline_create_infos;
     auto *crtpl_state = reinterpret_cast<create_ray_tracing_pipeline_api_state *>(crtpl_state_data);
-    PreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, crtpl_state->pipe_state, &new_pipeline_create_infos,
-                                   VK_PIPELINE_BIND_POINT_RAY_TRACING_NV);
+    SharedPreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, crtpl_state->pipe_state,
+                                         &new_pipeline_create_infos, VK_PIPELINE_BIND_POINT_RAY_TRACING_NV, this);
     crtpl_state->gpu_create_infos = new_pipeline_create_infos;
     crtpl_state->pCreateInfos = reinterpret_cast<VkRayTracingPipelineCreateInfoNV *>(crtpl_state->gpu_create_infos.data());
 }
@@ -1413,115 +1092,10 @@ void GpuAssisted::PreCallRecordCreateRayTracingPipelinesKHR(VkDevice device, VkP
                                                             void *crtpl_state_data) {
     std::vector<safe_VkRayTracingPipelineCreateInfoCommon> new_pipeline_create_infos;
     auto *crtpl_state = reinterpret_cast<create_ray_tracing_pipeline_khr_api_state *>(crtpl_state_data);
-    PreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, crtpl_state->pipe_state, &new_pipeline_create_infos,
-                                   VK_PIPELINE_BIND_POINT_RAY_TRACING_KHR);
+    SharedPreCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, crtpl_state->pipe_state,
+                                         &new_pipeline_create_infos, VK_PIPELINE_BIND_POINT_RAY_TRACING_KHR, this);
     crtpl_state->gpu_create_infos = new_pipeline_create_infos;
     crtpl_state->pCreateInfos = reinterpret_cast<VkRayTracingPipelineCreateInfoKHR *>(crtpl_state->gpu_create_infos.data());
-}
-template <typename CreateInfo>
-struct CreatePipelineTraits {};
-template <>
-struct CreatePipelineTraits<VkGraphicsPipelineCreateInfo> {
-    using SafeType = safe_VkGraphicsPipelineCreateInfo;
-    static const SafeType &GetPipelineCI(const PIPELINE_STATE *pipeline_state) { return pipeline_state->graphicsPipelineCI; }
-    static uint32_t GetStageCount(const VkGraphicsPipelineCreateInfo &createInfo) { return createInfo.stageCount; }
-    static VkShaderModule GetShaderModule(const VkGraphicsPipelineCreateInfo &createInfo, uint32_t stage) {
-        return createInfo.pStages[stage].module;
-    }
-    static void SetShaderModule(SafeType *createInfo, VkShaderModule shader_module, uint32_t stage) {
-        createInfo->pStages[stage].module = shader_module;
-    }
-};
-
-template <>
-struct CreatePipelineTraits<VkComputePipelineCreateInfo> {
-    using SafeType = safe_VkComputePipelineCreateInfo;
-    static const SafeType &GetPipelineCI(const PIPELINE_STATE *pipeline_state) { return pipeline_state->computePipelineCI; }
-    static uint32_t GetStageCount(const VkComputePipelineCreateInfo &createInfo) { return 1; }
-    static VkShaderModule GetShaderModule(const VkComputePipelineCreateInfo &createInfo, uint32_t stage) {
-        return createInfo.stage.module;
-    }
-    static void SetShaderModule(SafeType *createInfo, VkShaderModule shader_module, uint32_t stage) {
-        assert(stage == 0);
-        createInfo->stage.module = shader_module;
-    }
-};
-template <>
-struct CreatePipelineTraits<VkRayTracingPipelineCreateInfoNV> {
-    using SafeType = safe_VkRayTracingPipelineCreateInfoCommon;
-    static const SafeType &GetPipelineCI(const PIPELINE_STATE *pipeline_state) { return pipeline_state->raytracingPipelineCI; }
-    static uint32_t GetStageCount(const VkRayTracingPipelineCreateInfoNV &createInfo) { return createInfo.stageCount; }
-    static VkShaderModule GetShaderModule(const VkRayTracingPipelineCreateInfoNV &createInfo, uint32_t stage) {
-        return createInfo.pStages[stage].module;
-    }
-    static void SetShaderModule(SafeType *createInfo, VkShaderModule shader_module, uint32_t stage) {
-        createInfo->pStages[stage].module = shader_module;
-    }
-};
-template <>
-struct CreatePipelineTraits<VkRayTracingPipelineCreateInfoKHR> {
-    using SafeType = safe_VkRayTracingPipelineCreateInfoCommon;
-    static const SafeType &GetPipelineCI(const PIPELINE_STATE *pipeline_state) { return pipeline_state->raytracingPipelineCI; }
-    static uint32_t GetStageCount(const VkRayTracingPipelineCreateInfoKHR &createInfo) { return createInfo.stageCount; }
-    static VkShaderModule GetShaderModule(const VkRayTracingPipelineCreateInfoKHR &createInfo, uint32_t stage) {
-        return createInfo.pStages[stage].module;
-    }
-    static void SetShaderModule(SafeType *createInfo, VkShaderModule shader_module, uint32_t stage) {
-        createInfo->pStages[stage].module = shader_module;
-    }
-};
-
-// Examine the pipelines to see if they use the debug descriptor set binding index.
-// If any do, create new non-instrumented shader modules and use them to replace the instrumented
-// shaders in the pipeline.  Return the (possibly) modified create infos to the caller.
-template <typename CreateInfo, typename SafeCreateInfo>
-void GpuAssisted::PreCallRecordPipelineCreations(uint32_t count, const CreateInfo *pCreateInfos,
-                                                 const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                                 std::vector<std::shared_ptr<PIPELINE_STATE>> &pipe_state,
-                                                 std::vector<SafeCreateInfo> *new_pipeline_create_infos,
-                                                 const VkPipelineBindPoint bind_point) {
-    using Accessor = CreatePipelineTraits<CreateInfo>;
-    if (bind_point != VK_PIPELINE_BIND_POINT_GRAPHICS && bind_point != VK_PIPELINE_BIND_POINT_COMPUTE &&
-        bind_point != VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-        return;
-    }
-
-    // Walk through all the pipelines, make a copy of each and flag each pipeline that contains a shader that uses the debug
-    // descriptor set index.
-    for (uint32_t pipeline = 0; pipeline < count; ++pipeline) {
-        uint32_t stageCount = Accessor::GetStageCount(pCreateInfos[pipeline]);
-        new_pipeline_create_infos->push_back(Accessor::GetPipelineCI(pipe_state[pipeline].get()));
-
-        bool replace_shaders = false;
-        if (pipe_state[pipeline]->active_slots.find(desc_set_bind_index) != pipe_state[pipeline]->active_slots.end()) {
-            replace_shaders = true;
-        }
-        // If the app requests all available sets, the pipeline layout was not modified at pipeline layout creation and the already
-        // instrumented shaders need to be replaced with uninstrumented shaders
-        if (pipe_state[pipeline]->pipeline_layout->set_layouts.size() >= adjusted_max_desc_sets) {
-            replace_shaders = true;
-        }
-
-        if (replace_shaders) {
-            for (uint32_t stage = 0; stage < stageCount; ++stage) {
-                const SHADER_MODULE_STATE *shader = GetShaderModuleState(Accessor::GetShaderModule(pCreateInfos[pipeline], stage));
-
-                VkShaderModuleCreateInfo create_info = {};
-                VkShaderModule shader_module;
-                create_info.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-                create_info.pCode = shader->words.data();
-                create_info.codeSize = shader->words.size() * sizeof(uint32_t);
-                VkResult result = DispatchCreateShaderModule(device, &create_info, pAllocator, &shader_module);
-                if (result == VK_SUCCESS) {
-                    Accessor::SetShaderModule(&(*new_pipeline_create_infos)[pipeline], shader_module, stage);
-                } else {
-                    ReportSetupProblem(Accessor::GetShaderModule(pCreateInfos[pipeline], stage),
-                                       "Unable to replace instrumented shader with non-instrumented one.  "
-                                       "Device could become unstable.");
-                }
-            }
-        }
-    }
 }
 
 void GpuAssisted::PostCallRecordCreateGraphicsPipelines(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
@@ -1530,7 +1104,7 @@ void GpuAssisted::PostCallRecordCreateGraphicsPipelines(VkDevice device, VkPipel
                                                         VkResult result, void *cgpl_state_data) {
     ValidationStateTracker::PostCallRecordCreateGraphicsPipelines(device, pipelineCache, count, pCreateInfos, pAllocator,
                                                                   pPipelines, result, cgpl_state_data);
-    PostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_GRAPHICS);
+    SharedPostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_GRAPHICS, this);
 }
 
 void GpuAssisted::PostCallRecordCreateComputePipelines(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
@@ -1539,7 +1113,7 @@ void GpuAssisted::PostCallRecordCreateComputePipelines(VkDevice device, VkPipeli
                                                        VkResult result, void *ccpl_state_data) {
     ValidationStateTracker::PostCallRecordCreateComputePipelines(device, pipelineCache, count, pCreateInfos, pAllocator, pPipelines,
                                                                  result, ccpl_state_data);
-    PostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_COMPUTE);
+    SharedPostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_COMPUTE, this);
 }
 
 void GpuAssisted::PostCallRecordCreateRayTracingPipelinesNV(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
@@ -1548,7 +1122,7 @@ void GpuAssisted::PostCallRecordCreateRayTracingPipelinesNV(VkDevice device, VkP
                                                             VkResult result, void *crtpl_state_data) {
     ValidationStateTracker::PostCallRecordCreateRayTracingPipelinesNV(device, pipelineCache, count, pCreateInfos, pAllocator,
                                                                       pPipelines, result, crtpl_state_data);
-    PostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_RAY_TRACING_NV);
+    SharedPostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_RAY_TRACING_NV, this);
 }
 
 void GpuAssisted::PostCallRecordCreateRayTracingPipelinesKHR(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
@@ -1557,89 +1131,8 @@ void GpuAssisted::PostCallRecordCreateRayTracingPipelinesKHR(VkDevice device, Vk
                                                              VkResult result, void *crtpl_state_data) {
     ValidationStateTracker::PostCallRecordCreateRayTracingPipelinesKHR(device, pipelineCache, count, pCreateInfos, pAllocator,
                                                                        pPipelines, result, crtpl_state_data);
-    PostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_RAY_TRACING_KHR);
-}
-
-// For every pipeline:
-// - For every shader in a pipeline:
-//   - If the shader had to be replaced in PreCallRecord (because the pipeline is using the debug desc set index):
-//     - Destroy it since it has been bound into the pipeline by now.  This is our only chance to delete it.
-//   - Track the shader in the shader_map
-//   - Save the shader binary if it contains debug code
-template <typename CreateInfo>
-void GpuAssisted::PostCallRecordPipelineCreations(const uint32_t count, const CreateInfo *pCreateInfos,
-                                                  const VkAllocationCallbacks *pAllocator, VkPipeline *pPipelines,
-                                                  const VkPipelineBindPoint bind_point) {
-    using Accessor = CreatePipelineTraits<CreateInfo>;
-    if (bind_point != VK_PIPELINE_BIND_POINT_GRAPHICS && bind_point != VK_PIPELINE_BIND_POINT_COMPUTE &&
-        bind_point != VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-        return;
-    }
-    for (uint32_t pipeline = 0; pipeline < count; ++pipeline) {
-        auto pipeline_state = ValidationStateTracker::GetPipelineState(pPipelines[pipeline]);
-        if (nullptr == pipeline_state) continue;
-
-        uint32_t stageCount = 0;
-        if (bind_point == VK_PIPELINE_BIND_POINT_GRAPHICS) {
-            stageCount = pipeline_state->graphicsPipelineCI.stageCount;
-        } else if (bind_point == VK_PIPELINE_BIND_POINT_COMPUTE) {
-            stageCount = 1;
-        } else if (bind_point == VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-            stageCount = pipeline_state->raytracingPipelineCI.stageCount;
-        } else {
-            assert(false);
-        }
-
-        for (uint32_t stage = 0; stage < stageCount; ++stage) {
-            if (pipeline_state->active_slots.find(desc_set_bind_index) != pipeline_state->active_slots.end()) {
-                DispatchDestroyShaderModule(device, Accessor::GetShaderModule(pCreateInfos[pipeline], stage), pAllocator);
-            }
-
-            const SHADER_MODULE_STATE *shader_state = nullptr;
-            if (bind_point == VK_PIPELINE_BIND_POINT_GRAPHICS) {
-                shader_state = GetShaderModuleState(pipeline_state->graphicsPipelineCI.pStages[stage].module);
-            } else if (bind_point == VK_PIPELINE_BIND_POINT_COMPUTE) {
-                assert(stage == 0);
-                shader_state = GetShaderModuleState(pipeline_state->computePipelineCI.stage.module);
-            } else if (bind_point == VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-                shader_state = GetShaderModuleState(pipeline_state->raytracingPipelineCI.pStages[stage].module);
-            } else {
-                assert(false);
-            }
-
-            if (shader_state) {
-                std::vector<unsigned int> code;
-                // Save the shader binary if debug info is present.
-                // The core_validation ShaderModule tracker saves the binary too, but discards it when the ShaderModule
-                // is destroyed.  Applications may destroy ShaderModules after they are placed in a pipeline and before
-                // the pipeline is used, so we have to keep another copy.
-                if (shader_state->has_valid_spirv) {  // really checking for presense of SPIR-V code.
-                    for (auto insn : *shader_state) {
-                        if (insn.opcode() == spv::OpLine) {
-                            code = shader_state->words;
-                            break;
-                        }
-                    }
-                }
-                shader_map[shader_state->gpu_validation_shader_id].pipeline = pipeline_state->pipeline;
-                // Be careful to use the originally bound (instrumented) shader here, even if PreCallRecord had to back it
-                // out with a non-instrumented shader.  The non-instrumented shader (found in pCreateInfo) was destroyed above.
-                VkShaderModule shader_module = VK_NULL_HANDLE;
-                if (bind_point == VK_PIPELINE_BIND_POINT_GRAPHICS) {
-                    shader_module = pipeline_state->graphicsPipelineCI.pStages[stage].module;
-                } else if (bind_point == VK_PIPELINE_BIND_POINT_COMPUTE) {
-                    assert(stage == 0);
-                    shader_module = pipeline_state->computePipelineCI.stage.module;
-                } else if (bind_point == VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-                    shader_module = pipeline_state->raytracingPipelineCI.pStages[stage].module;
-                } else {
-                    assert(false);
-                }
-                shader_map[shader_state->gpu_validation_shader_id].shader_module = shader_module;
-                shader_map[shader_state->gpu_validation_shader_id].pgm = std::move(code);
-            }
-        }
-    }
+    SharedPostCallRecordPipelineCreations(count, pCreateInfos, pAllocator, pPipelines, VK_PIPELINE_BIND_POINT_RAY_TRACING_KHR,
+                                          this);
 }
 
 // Remove all the shader trackers associated with this destroyed pipeline.
@@ -1697,69 +1190,6 @@ void GpuAssisted::PreCallRecordCreateShaderModule(VkDevice device, const VkShade
     }
 }
 
-// Generate the stage-specific part of the message.
-static void GenerateStageMessage(const uint32_t *debug_record, std::string &msg) {
-    using namespace spvtools;
-    std::ostringstream strm;
-    switch (debug_record[kInstCommonOutStageIdx]) {
-        case spv::ExecutionModelVertex: {
-            strm << "Stage = Vertex. Vertex Index = " << debug_record[kInstVertOutVertexIndex]
-                 << " Instance Index = " << debug_record[kInstVertOutInstanceIndex] << ". ";
-        } break;
-        case spv::ExecutionModelTessellationControl: {
-            strm << "Stage = Tessellation Control.  Invocation ID = " << debug_record[kInstTessCtlOutInvocationId]
-                 << ", Primitive ID = " << debug_record[kInstTessCtlOutPrimitiveId];
-        } break;
-        case spv::ExecutionModelTessellationEvaluation: {
-            strm << "Stage = Tessellation Eval.  Primitive ID = " << debug_record[kInstTessEvalOutPrimitiveId]
-                 << ", TessCoord (u, v) = (" << debug_record[kInstTessEvalOutTessCoordU] << ", "
-                 << debug_record[kInstTessEvalOutTessCoordV] << "). ";
-        } break;
-        case spv::ExecutionModelGeometry: {
-            strm << "Stage = Geometry.  Primitive ID = " << debug_record[kInstGeomOutPrimitiveId]
-                 << " Invocation ID = " << debug_record[kInstGeomOutInvocationId] << ". ";
-        } break;
-        case spv::ExecutionModelFragment: {
-            strm << "Stage = Fragment.  Fragment coord (x,y) = ("
-                 << *reinterpret_cast<const float *>(&debug_record[kInstFragOutFragCoordX]) << ", "
-                 << *reinterpret_cast<const float *>(&debug_record[kInstFragOutFragCoordY]) << "). ";
-        } break;
-        case spv::ExecutionModelGLCompute: {
-            strm << "Stage = Compute.  Global invocation ID (x, y, z) = (" << debug_record[kInstCompOutGlobalInvocationIdX] << ", "
-                 << debug_record[kInstCompOutGlobalInvocationIdY] << ", " << debug_record[kInstCompOutGlobalInvocationIdZ] << " )";
-        } break;
-        case spv::ExecutionModelRayGenerationNV: {
-            strm << "Stage = Ray Generation.  Global Launch ID (x,y,z) = (" << debug_record[kInstRayTracingOutLaunchIdX] << ", "
-                 << debug_record[kInstRayTracingOutLaunchIdY] << ", " << debug_record[kInstRayTracingOutLaunchIdZ] << "). ";
-        } break;
-        case spv::ExecutionModelIntersectionNV: {
-            strm << "Stage = Intersection.  Global Launch ID (x,y,z) = (" << debug_record[kInstRayTracingOutLaunchIdX] << ", "
-                 << debug_record[kInstRayTracingOutLaunchIdY] << ", " << debug_record[kInstRayTracingOutLaunchIdZ] << "). ";
-        } break;
-        case spv::ExecutionModelAnyHitNV: {
-            strm << "Stage = Any Hit.  Global Launch ID (x,y,z) = (" << debug_record[kInstRayTracingOutLaunchIdX] << ", "
-                 << debug_record[kInstRayTracingOutLaunchIdY] << ", " << debug_record[kInstRayTracingOutLaunchIdZ] << "). ";
-        } break;
-        case spv::ExecutionModelClosestHitNV: {
-            strm << "Stage = Closest Hit.  Global Launch ID (x,y,z) = (" << debug_record[kInstRayTracingOutLaunchIdX] << ", "
-                 << debug_record[kInstRayTracingOutLaunchIdY] << ", " << debug_record[kInstRayTracingOutLaunchIdZ] << "). ";
-        } break;
-        case spv::ExecutionModelMissNV: {
-            strm << "Stage = Miss.  Global Launch ID (x,y,z) = (" << debug_record[kInstRayTracingOutLaunchIdX] << ", "
-                 << debug_record[kInstRayTracingOutLaunchIdY] << ", " << debug_record[kInstRayTracingOutLaunchIdZ] << "). ";
-        } break;
-        case spv::ExecutionModelCallableNV: {
-            strm << "Stage = Callable.  Global Launch ID (x,y,z) = (" << debug_record[kInstRayTracingOutLaunchIdX] << ", "
-                 << debug_record[kInstRayTracingOutLaunchIdY] << ", " << debug_record[kInstRayTracingOutLaunchIdZ] << "). ";
-        } break;
-        default: {
-            strm << "Internal Error (unexpected stage = " << debug_record[kInstCommonOutStageIdx] << "). ";
-            assert(false);
-        } break;
-    }
-    msg = strm.str();
-}
-
 // Generate the part of the message describing the violation.
 static void GenerateValidationMessage(const uint32_t *debug_record, std::string &msg, std::string &vuid_msg) {
     using namespace spvtools;
@@ -1788,260 +1218,6 @@ static void GenerateValidationMessage(const uint32_t *debug_record, std::string 
     msg = strm.str();
 }
 
-static std::string LookupDebugUtilsName(const debug_report_data *report_data, const uint64_t object) {
-    auto object_label = report_data->DebugReportGetUtilsObjectName(object);
-    if (object_label != "") {
-        object_label = "(" + object_label + ")";
-    }
-    return object_label;
-}
-
-// Generate message from the common portion of the debug report record.
-static void GenerateCommonMessage(const debug_report_data *report_data, const CMD_BUFFER_STATE *cb_node,
-                                  const uint32_t *debug_record, const VkShaderModule shader_module_handle,
-                                  const VkPipeline pipeline_handle, const VkPipelineBindPoint pipeline_bind_point,
-                                  const uint32_t operation_index, std::string &msg) {
-    using namespace spvtools;
-    std::ostringstream strm;
-    if (shader_module_handle == VK_NULL_HANDLE) {
-        strm << std::hex << std::showbase << "Internal Error: Unable to locate information for shader used in command buffer "
-             << LookupDebugUtilsName(report_data, HandleToUint64(cb_node->commandBuffer)) << "("
-             << HandleToUint64(cb_node->commandBuffer) << "). ";
-        assert(true);
-    } else {
-        strm << std::hex << std::showbase << "Command buffer "
-             << LookupDebugUtilsName(report_data, HandleToUint64(cb_node->commandBuffer)) << "("
-             << HandleToUint64(cb_node->commandBuffer) << "). ";
-        if (pipeline_bind_point == VK_PIPELINE_BIND_POINT_GRAPHICS) {
-            strm << "Draw ";
-        } else if (pipeline_bind_point == VK_PIPELINE_BIND_POINT_COMPUTE) {
-            strm << "Compute ";
-        } else if (pipeline_bind_point == VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-            strm << "Ray Trace ";
-        } else {
-            assert(false);
-            strm << "Unknown Pipeline Operation ";
-        }
-        strm << "Index " << operation_index << ". "
-             << "Pipeline " << LookupDebugUtilsName(report_data, HandleToUint64(pipeline_handle)) << "("
-             << HandleToUint64(pipeline_handle) << "). "
-             << "Shader Module " << LookupDebugUtilsName(report_data, HandleToUint64(shader_module_handle)) << "("
-             << HandleToUint64(shader_module_handle) << "). ";
-    }
-    strm << std::dec << std::noshowbase;
-    strm << "Shader Instruction Index = " << debug_record[kInstCommonOutInstructionIdx] << ". ";
-    msg = strm.str();
-}
-
-// Read the contents of the SPIR-V OpSource instruction and any following continuation instructions.
-// Split the single string into a vector of strings, one for each line, for easier processing.
-static void ReadOpSource(const SHADER_MODULE_STATE &shader, const uint32_t reported_file_id,
-                         std::vector<std::string> &opsource_lines) {
-    for (auto insn : shader) {
-        if ((insn.opcode() == spv::OpSource) && (insn.len() >= 5) && (insn.word(3) == reported_file_id)) {
-            std::istringstream in_stream;
-            std::string cur_line;
-            in_stream.str((char *)&insn.word(4));
-            while (std::getline(in_stream, cur_line)) {
-                opsource_lines.push_back(cur_line);
-            }
-            while ((++insn).opcode() == spv::OpSourceContinued) {
-                in_stream.str((char *)&insn.word(1));
-                while (std::getline(in_stream, cur_line)) {
-                    opsource_lines.push_back(cur_line);
-                }
-            }
-            break;
-        }
-    }
-}
-
-// The task here is to search the OpSource content to find the #line directive with the
-// line number that is closest to, but still prior to the reported error line number and
-// still within the reported filename.
-// From this known position in the OpSource content we can add the difference between
-// the #line line number and the reported error line number to determine the location
-// in the OpSource content of the reported error line.
-//
-// Considerations:
-// - Look only at #line directives that specify the reported_filename since
-//   the reported error line number refers to its location in the reported filename.
-// - If a #line directive does not have a filename, the file is the reported filename, or
-//   the filename found in a prior #line directive.  (This is C-preprocessor behavior)
-// - It is possible (e.g., inlining) for blocks of code to get shuffled out of their
-//   original order and the #line directives are used to keep the numbering correct.  This
-//   is why we need to examine the entire contents of the source, instead of leaving early
-//   when finding a #line line number larger than the reported error line number.
-//
-
-// GCC 4.8 has a problem with std::regex that is fixed in GCC 4.9.  Provide fallback code for 4.8
-#define GCC_VERSION (__GNUC__ * 10000 + __GNUC_MINOR__ * 100 + __GNUC_PATCHLEVEL__)
-
-#if defined(__GNUC__) && GCC_VERSION < 40900
-static bool GetLineAndFilename(const std::string string, uint32_t *linenumber, std::string &filename) {
-    // # line <linenumber> "<filename>" or
-    // #line <linenumber> "<filename>"
-    std::vector<std::string> tokens;
-    std::stringstream stream(string);
-    std::string temp;
-    uint32_t line_index = 0;
-
-    while (stream >> temp) tokens.push_back(temp);
-    auto size = tokens.size();
-    if (size > 1) {
-        if (tokens[0] == "#" && tokens[1] == "line") {
-            line_index = 2;
-        } else if (tokens[0] == "#line") {
-            line_index = 1;
-        }
-    }
-    if (0 == line_index) return false;
-    *linenumber = std::stoul(tokens[line_index]);
-    uint32_t filename_index = line_index + 1;
-    // Remove enclosing double quotes around filename
-    if (size > filename_index) filename = tokens[filename_index].substr(1, tokens[filename_index].size() - 2);
-    return true;
-}
-#else
-static bool GetLineAndFilename(const std::string string, uint32_t *linenumber, std::string &filename) {
-    static const std::regex line_regex(  // matches #line directives
-        "^"                              // beginning of line
-        "\\s*"                           // optional whitespace
-        "#"                              // required text
-        "\\s*"                           // optional whitespace
-        "line"                           // required text
-        "\\s+"                           // required whitespace
-        "([0-9]+)"                       // required first capture - line number
-        "(\\s+)?"                        // optional second capture - whitespace
-        "(\".+\")?"                      // optional third capture - quoted filename with at least one char inside
-        ".*");                           // rest of line (needed when using std::regex_match since the entire line is tested)
-
-    std::smatch captures;
-
-    bool found_line = std::regex_match(string, captures, line_regex);
-    if (!found_line) return false;
-
-    // filename is optional and considered found only if the whitespace and the filename are captured
-    if (captures[2].matched && captures[3].matched) {
-        // Remove enclosing double quotes.  The regex guarantees the quotes and at least one char.
-        filename = captures[3].str().substr(1, captures[3].str().size() - 2);
-    }
-    *linenumber = std::stoul(captures[1]);
-    return true;
-}
-#endif  // GCC_VERSION
-
-// Extract the filename, line number, and column number from the correct OpLine and build a message string from it.
-// Scan the source (from OpSource) to find the line of source at the reported line number and place it in another message string.
-static void GenerateSourceMessages(const std::vector<unsigned int> &pgm, const uint32_t *debug_record, std::string &filename_msg,
-                                   std::string &source_msg) {
-    using namespace spvtools;
-    std::ostringstream filename_stream;
-    std::ostringstream source_stream;
-    SHADER_MODULE_STATE shader;
-    shader.words = pgm;
-    // Find the OpLine just before the failing instruction indicated by the debug info.
-    // SPIR-V can only be iterated in the forward direction due to its opcode/length encoding.
-    uint32_t instruction_index = 0;
-    uint32_t reported_file_id = 0;
-    uint32_t reported_line_number = 0;
-    uint32_t reported_column_number = 0;
-    if (shader.words.size() > 0) {
-        for (auto insn : shader) {
-            if (insn.opcode() == spv::OpLine) {
-                reported_file_id = insn.word(1);
-                reported_line_number = insn.word(2);
-                reported_column_number = insn.word(3);
-            }
-            if (instruction_index == debug_record[kInstCommonOutInstructionIdx]) {
-                break;
-            }
-            instruction_index++;
-        }
-    }
-    // Create message with file information obtained from the OpString pointed to by the discovered OpLine.
-    std::string reported_filename;
-    if (reported_file_id == 0) {
-        filename_stream
-            << "Unable to find SPIR-V OpLine for source information.  Build shader with debug info to get source information.";
-    } else {
-        bool found_opstring = false;
-        for (auto insn : shader) {
-            if ((insn.opcode() == spv::OpString) && (insn.len() >= 3) && (insn.word(1) == reported_file_id)) {
-                found_opstring = true;
-                reported_filename = (char *)&insn.word(2);
-                if (reported_filename.empty()) {
-                    filename_stream << "Shader validation error occurred at line " << reported_line_number;
-                } else {
-                    filename_stream << "Shader validation error occurred in file: " << reported_filename << " at line "
-                                    << reported_line_number;
-                }
-                if (reported_column_number > 0) {
-                    filename_stream << ", column " << reported_column_number;
-                }
-                filename_stream << ".";
-                break;
-            }
-        }
-        if (!found_opstring) {
-            filename_stream << "Unable to find SPIR-V OpString for file id " << reported_file_id << " from OpLine instruction.";
-        }
-    }
-    filename_msg = filename_stream.str();
-
-    // Create message to display source code line containing error.
-    if ((reported_file_id != 0)) {
-        // Read the source code and split it up into separate lines.
-        std::vector<std::string> opsource_lines;
-        ReadOpSource(shader, reported_file_id, opsource_lines);
-        // Find the line in the OpSource content that corresponds to the reported error file and line.
-        if (!opsource_lines.empty()) {
-            uint32_t saved_line_number = 0;
-            std::string current_filename = reported_filename;  // current "preprocessor" filename state.
-            std::vector<std::string>::size_type saved_opsource_offset = 0;
-            bool found_best_line = false;
-            for (auto it = opsource_lines.begin(); it != opsource_lines.end(); ++it) {
-                uint32_t parsed_line_number;
-                std::string parsed_filename;
-                bool found_line = GetLineAndFilename(*it, &parsed_line_number, parsed_filename);
-                if (!found_line) continue;
-
-                bool found_filename = parsed_filename.size() > 0;
-                if (found_filename) {
-                    current_filename = parsed_filename;
-                }
-                if ((!found_filename) || (current_filename == reported_filename)) {
-                    // Update the candidate best line directive, if the current one is prior and closer to the reported line
-                    if (reported_line_number >= parsed_line_number) {
-                        if (!found_best_line ||
-                            (reported_line_number - parsed_line_number <= reported_line_number - saved_line_number)) {
-                            saved_line_number = parsed_line_number;
-                            saved_opsource_offset = std::distance(opsource_lines.begin(), it);
-                            found_best_line = true;
-                        }
-                    }
-                }
-            }
-            if (found_best_line) {
-                assert(reported_line_number >= saved_line_number);
-                std::vector<std::string>::size_type opsource_index =
-                    (reported_line_number - saved_line_number) + 1 + saved_opsource_offset;
-                if (opsource_index < opsource_lines.size()) {
-                    source_stream << "\n" << reported_line_number << ": " << opsource_lines[opsource_index].c_str();
-                } else {
-                    source_stream << "Internal error: calculated source line of " << opsource_index << " for source size of "
-                                  << opsource_lines.size() << " lines.";
-                }
-            } else {
-                source_stream << "Unable to find suitable #line directive in SPIR-V OpSource.";
-            }
-        } else {
-            source_stream << "Unable to find SPIR-V OpSource.";
-        }
-    }
-    source_msg = source_stream.str();
-}
-
 // Pull together all the information from the debug record to build the error message strings,
 // and then assemble them into a single message string.
 // Retrieve the shader program referenced by the unique shader ID provided in the debug record.
@@ -2049,8 +1225,8 @@ static void GenerateSourceMessages(const std::vector<unsigned int> &pgm, const u
 // sure it is available when the pipeline is submitted.  (The ShaderModule tracking object also
 // keeps a copy, but it can be destroyed after the pipeline is created and before it is submitted.)
 //
-void GpuAssisted::AnalyzeAndReportError(CMD_BUFFER_STATE *cb_node, VkQueue queue, VkPipelineBindPoint pipeline_bind_point,
-                                        uint32_t operation_index, uint32_t *const debug_output_buffer) {
+void GpuAssisted::AnalyzeAndGenerateMessages(VkCommandBuffer command_buffer, VkQueue queue, VkPipelineBindPoint pipeline_bind_point,
+                                             uint32_t operation_index, uint32_t *const debug_output_buffer) {
     using namespace spvtools;
     const uint32_t total_words = debug_output_buffer[0];
     // A zero here means that the shader instrumentation didn't write anything.
@@ -2087,10 +1263,10 @@ void GpuAssisted::AnalyzeAndReportError(CMD_BUFFER_STATE *cb_node, VkQueue queue
         pgm = it->second.pgm;
     }
     GenerateValidationMessage(debug_record, validation_message, vuid_msg);
-    GenerateStageMessage(debug_record, stage_message);
-    GenerateCommonMessage(report_data, cb_node, debug_record, shader_module_handle, pipeline_handle, pipeline_bind_point,
-                          operation_index, common_message);
-    GenerateSourceMessages(pgm, debug_record, filename_message, source_message);
+    SharedGenerateStageMessage(debug_record, stage_message);
+    SharedGenerateCommonMessage(report_data, command_buffer, debug_record, shader_module_handle, pipeline_handle,
+                                pipeline_bind_point, operation_index, common_message);
+    SharedGenerateSourceMessages(pgm, debug_record, false, filename_message, source_message);
     LogError(queue, vuid_msg.c_str(), "%s %s %s %s%s", validation_message.c_str(), common_message.c_str(), stage_message.c_str(),
              filename_message.c_str(), source_message.c_str());
     // The debug record at word kInstCommonOutSize is the number of words in the record
@@ -2099,50 +1275,9 @@ void GpuAssisted::AnalyzeAndReportError(CMD_BUFFER_STATE *cb_node, VkQueue queue
     memset(debug_output_buffer, 0, sizeof(uint32_t) * words_to_clear);
 }
 
-// For the given command buffer, map its debug data buffers and read their contents for analysis.
-void GpuAssisted::ProcessInstrumentationBuffer(VkQueue queue, CMD_BUFFER_STATE *cb_node) {
-    if (cb_node && (cb_node->hasDrawCmd || cb_node->hasTraceRaysCmd || cb_node->hasDispatchCmd)) {
-        auto gpu_buffer_list = GetGpuAssistedBufferInfo(cb_node->commandBuffer);
-        uint32_t draw_index = 0;
-        uint32_t compute_index = 0;
-        uint32_t ray_trace_index = 0;
-
-        for (auto &buffer_info : gpu_buffer_list) {
-            char *pData;
-            VkResult result = vmaMapMemory(vmaAllocator, buffer_info.output_mem_block.allocation, (void **)&pData);
-            // Analyze debug output buffer
-            if (result == VK_SUCCESS) {
-                uint32_t operation_index = 0;
-                if (buffer_info.pipeline_bind_point == VK_PIPELINE_BIND_POINT_GRAPHICS) {
-                    operation_index = draw_index;
-                } else if (buffer_info.pipeline_bind_point == VK_PIPELINE_BIND_POINT_COMPUTE) {
-                    operation_index = compute_index;
-                } else if (buffer_info.pipeline_bind_point == VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-                    operation_index = ray_trace_index;
-                } else {
-                    assert(false);
-                }
-
-                AnalyzeAndReportError(cb_node, queue, buffer_info.pipeline_bind_point, operation_index, (uint32_t *)pData);
-                vmaUnmapMemory(vmaAllocator, buffer_info.output_mem_block.allocation);
-            }
-
-            if (buffer_info.pipeline_bind_point == VK_PIPELINE_BIND_POINT_GRAPHICS) {
-                draw_index++;
-            } else if (buffer_info.pipeline_bind_point == VK_PIPELINE_BIND_POINT_COMPUTE) {
-                compute_index++;
-            } else if (buffer_info.pipeline_bind_point == VK_PIPELINE_BIND_POINT_RAY_TRACING_NV) {
-                ray_trace_index++;
-            } else {
-                assert(false);
-            }
-        }
-    }
-}
-
 // For the given command buffer, map its debug data buffers and update the status of any update after bind descriptors
 void GpuAssisted::UpdateInstrumentationBuffer(CMD_BUFFER_STATE *cb_node) {
-    auto gpu_buffer_list = GetGpuAssistedBufferInfo(cb_node->commandBuffer);
+    auto gpu_buffer_list = GetBufferInfo(cb_node->commandBuffer);
     uint32_t *pData;
     for (auto &buffer_info : gpu_buffer_list) {
         if (buffer_info.di_input_mem_block.update_at_submit.size() > 0) {
@@ -2154,75 +1289,6 @@ void GpuAssisted::UpdateInstrumentationBuffer(CMD_BUFFER_STATE *cb_node) {
                 vmaUnmapMemory(vmaAllocator, buffer_info.di_input_mem_block.allocation);
             }
         }
-    }
-}
-
-// Submit a memory barrier on graphics queues.
-// Lazy-create and record the needed command buffer.
-void GpuAssisted::SubmitBarrier(VkQueue queue) {
-    auto queue_barrier_command_info_it = queue_barrier_command_infos.emplace(queue, GpuAssistedQueueBarrierCommandInfo{});
-    if (queue_barrier_command_info_it.second) {
-        GpuAssistedQueueBarrierCommandInfo &queue_barrier_command_info = queue_barrier_command_info_it.first->second;
-
-        uint32_t queue_family_index = 0;
-
-        auto queue_state_it = queueMap.find(queue);
-        if (queue_state_it != queueMap.end()) {
-            queue_family_index = queue_state_it->second.queueFamilyIndex;
-        }
-
-        VkResult result = VK_SUCCESS;
-
-        VkCommandPoolCreateInfo pool_create_info = {};
-        pool_create_info.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-        pool_create_info.queueFamilyIndex = queue_family_index;
-        result = DispatchCreateCommandPool(device, &pool_create_info, nullptr, &queue_barrier_command_info.barrier_command_pool);
-        if (result != VK_SUCCESS) {
-            ReportSetupProblem(device, "Unable to create command pool for barrier CB.");
-            queue_barrier_command_info.barrier_command_pool = VK_NULL_HANDLE;
-            return;
-        }
-
-        VkCommandBufferAllocateInfo buffer_alloc_info = {};
-        buffer_alloc_info.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-        buffer_alloc_info.commandPool = queue_barrier_command_info.barrier_command_pool;
-        buffer_alloc_info.commandBufferCount = 1;
-        buffer_alloc_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-        result = DispatchAllocateCommandBuffers(device, &buffer_alloc_info, &queue_barrier_command_info.barrier_command_buffer);
-        if (result != VK_SUCCESS) {
-            ReportSetupProblem(device, "Unable to create barrier command buffer.");
-            DispatchDestroyCommandPool(device, queue_barrier_command_info.barrier_command_pool, nullptr);
-            queue_barrier_command_info.barrier_command_pool = VK_NULL_HANDLE;
-            queue_barrier_command_info.barrier_command_buffer = VK_NULL_HANDLE;
-            return;
-        }
-
-        // Hook up command buffer dispatch
-        vkSetDeviceLoaderData(device, queue_barrier_command_info.barrier_command_buffer);
-
-        // Record a global memory barrier to force availability of device memory operations to the host domain.
-        VkCommandBufferBeginInfo command_buffer_begin_info = {};
-        command_buffer_begin_info.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-        result = DispatchBeginCommandBuffer(queue_barrier_command_info.barrier_command_buffer, &command_buffer_begin_info);
-        if (result == VK_SUCCESS) {
-            VkMemoryBarrier memory_barrier = {};
-            memory_barrier.sType = VK_STRUCTURE_TYPE_MEMORY_BARRIER;
-            memory_barrier.srcAccessMask = VK_ACCESS_MEMORY_WRITE_BIT;
-            memory_barrier.dstAccessMask = VK_ACCESS_HOST_READ_BIT;
-
-            DispatchCmdPipelineBarrier(queue_barrier_command_info.barrier_command_buffer, VK_PIPELINE_STAGE_ALL_COMMANDS_BIT,
-                                       VK_PIPELINE_STAGE_HOST_BIT, 0, 1, &memory_barrier, 0, nullptr, 0, nullptr);
-            DispatchEndCommandBuffer(queue_barrier_command_info.barrier_command_buffer);
-        }
-    }
-
-    GpuAssistedQueueBarrierCommandInfo &queue_barrier_command_info = queue_barrier_command_info_it.first->second;
-    if (queue_barrier_command_info.barrier_command_buffer != VK_NULL_HANDLE) {
-        VkSubmitInfo submit_info = {};
-        submit_info.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-        submit_info.commandBufferCount = 1;
-        submit_info.pCommandBuffers = &queue_barrier_command_info.barrier_command_buffer;
-        DispatchQueueSubmit(queue, 1, &submit_info, VK_NULL_HANDLE);
     }
 }
 
@@ -2253,17 +1319,16 @@ void GpuAssisted::PostCallRecordQueueSubmit(VkQueue queue, uint32_t submitCount,
         const VkSubmitInfo *submit = &pSubmits[submit_idx];
         for (uint32_t i = 0; i < submit->commandBufferCount; i++) {
             auto cb_node = GetCBState(submit->pCommandBuffers[i]);
-            if (GetGpuAssistedBufferInfo(cb_node->commandBuffer).size() || cb_node->hasBuildAccelerationStructureCmd)
-                buffers_present = true;
+            if (GetBufferInfo(cb_node->commandBuffer).size() || cb_node->hasBuildAccelerationStructureCmd) buffers_present = true;
             for (auto secondaryCmdBuffer : cb_node->linkedCommandBuffers) {
-                if (GetGpuAssistedBufferInfo(secondaryCmdBuffer->commandBuffer).size() || cb_node->hasBuildAccelerationStructureCmd)
+                if (GetBufferInfo(secondaryCmdBuffer->commandBuffer).size() || cb_node->hasBuildAccelerationStructureCmd)
                     buffers_present = true;
             }
         }
     }
     if (!buffers_present) return;
 
-    SubmitBarrier(queue);
+    SharedSubmitBarrier(queue, this);
 
     DispatchQueueWaitIdle(queue);
 
@@ -2271,10 +1336,10 @@ void GpuAssisted::PostCallRecordQueueSubmit(VkQueue queue, uint32_t submitCount,
         const VkSubmitInfo *submit = &pSubmits[submit_idx];
         for (uint32_t i = 0; i < submit->commandBufferCount; i++) {
             auto cb_node = GetCBState(submit->pCommandBuffers[i]);
-            ProcessInstrumentationBuffer(queue, cb_node);
+            SharedProcessInstrumentationBuffer(queue, cb_node, this);
             ProcessAccelerationStructureBuildValidationBuffer(queue, cb_node);
             for (auto secondaryCmdBuffer : cb_node->linkedCommandBuffers) {
-                ProcessInstrumentationBuffer(queue, secondaryCmdBuffer);
+                SharedProcessInstrumentationBuffer(queue, secondaryCmdBuffer, this);
                 ProcessAccelerationStructureBuildValidationBuffer(queue, cb_node);
             }
         }
@@ -2379,7 +1444,7 @@ void GpuAssisted::AllocateValidationResources(const VkCommandBuffer cmd_buffer, 
 
     std::vector<VkDescriptorSet> desc_sets;
     VkDescriptorPool desc_pool = VK_NULL_HANDLE;
-    result = desc_set_manager->GetDescriptorSets(1, &desc_pool, &desc_sets);
+    result = desc_set_manager->GetDescriptorSets(1, &desc_pool, debug_desc_layout, &desc_sets);
     assert(result == VK_SUCCESS);
     if (result != VK_SUCCESS) {
         ReportSetupProblem(device, "Unable to allocate descriptor sets.  Device could become unstable.");
@@ -2630,8 +1695,7 @@ void GpuAssisted::AllocateValidationResources(const VkCommandBuffer cmd_buffer, 
                                           desc_sets.data(), 0, nullptr);
         }
         // Record buffer and memory info in CB state tracking
-        GetGpuAssistedBufferInfo(cmd_buffer)
-            .emplace_back(output_block, di_input_block, bda_input_block, desc_sets[0], desc_pool, bind_point);
+        GetBufferInfo(cmd_buffer).emplace_back(output_block, di_input_block, bda_input_block, desc_sets[0], desc_pool, bind_point);
     } else {
         ReportSetupProblem(device, "Unable to find pipeline state");
         vmaDestroyBuffer(vmaAllocator, di_input_block.buffer, di_input_block.allocation);
