+++ layers/shader_debug_printf.h
@@ -22,6 +22,7 @@
 #include "chassis.h"
 #include "vk_mem_alloc.h"
 #include "state_tracker.h"
+#include "gpu_shared.h"
 #include <map>
 class ShaderPrintf;
 
@@ -38,32 +39,6 @@ struct SPFBufferInfo {
     SPFBufferInfo(SPFDeviceMemoryBlock output_mem_block, VkDescriptorSet desc_set, VkDescriptorPool desc_pool,
                   VkPipelineBindPoint pipeline_bind_point)
         : output_mem_block(output_mem_block), desc_set(desc_set), desc_pool(desc_pool), pipeline_bind_point(pipeline_bind_point){};
-};
-struct SPFQueueBarrierCommandInfo {
-    VkCommandPool barrier_command_pool = VK_NULL_HANDLE;
-    VkCommandBuffer barrier_command_buffer = VK_NULL_HANDLE;
-};
-
-// Class to encapsulate Descriptor Set allocation.  This manager creates and destroys Descriptor Pools
-// as needed to satisfy requests for descriptor sets.
-class SPFDescriptorSetManager {
-  public:
-    SPFDescriptorSetManager(ShaderPrintf* dev_data);
-    ~SPFDescriptorSetManager();
-
-    VkResult GetDescriptorSet(VkDescriptorPool* desc_pool, VkDescriptorSet* desc_sets);
-    VkResult GetDescriptorSets(uint32_t count, VkDescriptorPool* pool, std::vector<VkDescriptorSet>* desc_sets);
-    void PutBackDescriptorSet(VkDescriptorPool desc_pool, VkDescriptorSet desc_set);
-
-  private:
-    static const uint32_t kItemsPerChunk = 512;
-    struct PoolTracker {
-        uint32_t size;
-        uint32_t used;
-    };
-
-    ShaderPrintf* dev_data_;
-    std::unordered_map<VkDescriptorPool, struct PoolTracker> desc_pool_map_;
 };
 
 struct SPFShaderTracker {
@@ -93,19 +68,26 @@ struct SPFOutputRecord {
 };
 
 class ShaderPrintf : public ValidationStateTracker {
-    bool aborted = false;
     VkPhysicalDeviceFeatures supported_features;
-    uint32_t adjusted_max_desc_sets;
-    uint32_t desc_set_bind_index;
+
     uint32_t unique_shader_module_id = 0;
-    std::unordered_map<uint32_t, SPFShaderTracker> shader_map;
-    std::unique_ptr<SPFDescriptorSetManager> desc_set_manager;
-    std::map<VkQueue, SPFQueueBarrierCommandInfo> queue_barrier_command_infos;
     std::unordered_map<VkCommandBuffer, std::vector<SPFBufferInfo>> command_buffer_map;
     uint32_t output_buffer_size;
-    VmaAllocator vmaAllocator = {};
+
+  public:
+    bool aborted = false;
+    VkDevice device;
+    VkPhysicalDevice physicalDevice;
+    uint32_t adjusted_max_desc_sets;
+    uint32_t desc_set_bind_index;
+    VkDescriptorSetLayout debug_desc_layout;
+    VkDescriptorSetLayout dummy_desc_layout;
+    std::unique_ptr<SharedDescriptorSetManager> desc_set_manager;
+    std::unordered_map<uint32_t, SPFShaderTracker> shader_map;
     PFN_vkSetDeviceLoaderData vkSetDeviceLoaderData;
-    std::vector<SPFBufferInfo>& GetSPFBufferInfo(const VkCommandBuffer command_buffer) {
+    VmaAllocator vmaAllocator = {};
+    std::map<VkQueue, SharedQueueBarrierCommandInfo> queue_barrier_command_infos;
+    std::vector<SPFBufferInfo>& GetBufferInfo(const VkCommandBuffer command_buffer) {
         auto buffer_list = command_buffer_map.find(command_buffer);
         if (buffer_list == command_buffer_map.end()) {
             std::vector<SPFBufferInfo> new_list{};
@@ -114,17 +96,15 @@ class ShaderPrintf : public ValidationStateTracker {
         }
         return buffer_list->second;
     }
-    void ReportSetupProblem(const char* const specific_message) const;
 
-  public:
-    VkDescriptorSetLayout debug_desc_layout;
-    VkDescriptorSetLayout dummy_desc_layout;
+    template <typename T>
+    void ReportSetupProblem(T object, const char* const specific_message) const;
     void PreCallRecordCreateDevice(VkPhysicalDevice gpu, const VkDeviceCreateInfo* pCreateInfo,
                                    const VkAllocationCallbacks* pAllocator, VkDevice* pDevice,
                                    safe_VkDeviceCreateInfo* modified_create_info);
     void PostCallRecordCreateDevice(VkPhysicalDevice gpu, const VkDeviceCreateInfo* pCreateInfo,
                                     const VkAllocationCallbacks* pAllocator, VkDevice* pDevice, VkResult result);
-    void ShaderPrintf::PreCallRecordDestroyDevice(VkDevice device, const VkAllocationCallbacks* pAllocator);
+    void PreCallRecordDestroyDevice(VkDevice device, const VkAllocationCallbacks* pAllocator);
     void PreCallRecordCreatePipelineLayout(VkDevice device, const VkPipelineLayoutCreateInfo* pCreateInfo,
                                            const VkAllocationCallbacks* pAllocator, VkPipelineLayout* pPipelineLayout,
                                            void* cpl_state_data);
@@ -153,11 +133,6 @@ class ShaderPrintf : public ValidationStateTracker {
                                                    const VkRayTracingPipelineCreateInfoKHR* pCreateInfos,
                                                    const VkAllocationCallbacks* pAllocator, VkPipeline* pPipelines,
                                                    void* crtpl_state_data);
-    template <typename CreateInfo, typename SafeCreateInfo>
-    void PreCallRecordPipelineCreations(uint32_t count, const CreateInfo* pCreateInfos, const VkAllocationCallbacks* pAllocator,
-                                        VkPipeline* pPipelines, std::vector<std::shared_ptr<PIPELINE_STATE>>& pipe_state,
-                                        std::vector<SafeCreateInfo>* new_pipeline_create_infos,
-                                        const VkPipelineBindPoint bind_point);
     void PostCallRecordCreateGraphicsPipelines(VkDevice device, VkPipelineCache pipelineCache, uint32_t count,
                                                const VkGraphicsPipelineCreateInfo* pCreateInfos,
                                                const VkAllocationCallbacks* pAllocator, VkPipeline* pPipelines, VkResult result,
@@ -170,10 +145,6 @@ class ShaderPrintf : public ValidationStateTracker {
                                                    const VkRayTracingPipelineCreateInfoNV* pCreateInfos,
                                                    const VkAllocationCallbacks* pAllocator, VkPipeline* pPipelines, VkResult result,
                                                    void* crtpl_state_data);
-    template <typename CreateInfo>
-    void PostCallRecordPipelineCreations(const uint32_t count, const CreateInfo* pCreateInfos,
-                                         const VkAllocationCallbacks* pAllocator, VkPipeline* pPipelines,
-                                         const VkPipelineBindPoint bind_point);
     void PreCallRecordDestroyPipeline(VkDevice device, VkPipeline pipeline, const VkAllocationCallbacks* pAllocator);
     bool InstrumentShader(const VkShaderModuleCreateInfo* pCreateInfo, std::vector<unsigned int>& new_pgm,
                           uint32_t* unique_shader_id);
@@ -182,10 +153,8 @@ class ShaderPrintf : public ValidationStateTracker {
                                          void* csm_state_data);
     std::vector<SPFSubstring> ParseFormatString(std::string format_string);
     std::string FindFormatString(std::vector<unsigned int> pgm, uint32_t string_id);
-    void AnalyzeAndPrintfMessages(CMD_BUFFER_STATE* cb_node, VkQueue queue, VkPipelineBindPoint pipeline_bind_point,
-                                  uint32_t operation_index, uint32_t* const debug_output_buffer);
-    void ProcessInstrumentationBuffer(VkQueue queue, CMD_BUFFER_STATE* cb_node);
-    void SubmitBarrier(VkQueue queue);
+    void AnalyzeAndGenerateMessages(VkCommandBuffer command_buffer, VkQueue queue, VkPipelineBindPoint pipeline_bind_point,
+                                    uint32_t operation_index, uint32_t* const debug_output_buffer);
     void PreCallRecordCmdDraw(VkCommandBuffer commandBuffer, uint32_t vertexCount, uint32_t instanceCount, uint32_t firstVertex,
                               uint32_t firstInstance);
     void PreCallRecordCmdDrawIndexed(VkCommandBuffer commandBuffer, uint32_t indexCount, uint32_t instanceCount,
@@ -210,8 +179,29 @@ class ShaderPrintf : public ValidationStateTracker {
                                       VkDeviceSize hitShaderBindingStride, VkBuffer callableShaderBindingTableBuffer,
                                       VkDeviceSize callableShaderBindingOffset, VkDeviceSize callableShaderBindingStride,
                                       uint32_t width, uint32_t height, uint32_t depth);
+    void PreCallRecordCmdTraceRaysKHR(VkCommandBuffer commandBuffer, const VkStridedBufferRegionKHR* pRaygenShaderBindingTable,
+                                      const VkStridedBufferRegionKHR* pMissShaderBindingTable,
+                                      const VkStridedBufferRegionKHR* pHitShaderBindingTable,
+                                      const VkStridedBufferRegionKHR* pCallableShaderBindingTable, uint32_t width, uint32_t height,
+                                      uint32_t depth);
+    void PostCallRecordCmdTraceRaysKHR(VkCommandBuffer commandBuffer, const VkStridedBufferRegionKHR* pRaygenShaderBindingTable,
+                                       const VkStridedBufferRegionKHR* pMissShaderBindingTable,
+                                       const VkStridedBufferRegionKHR* pHitShaderBindingTable,
+                                       const VkStridedBufferRegionKHR* pCallableShaderBindingTable, uint32_t width, uint32_t height,
+                                       uint32_t depth);
+    void PreCallRecordCmdTraceRaysIndirectKHR(VkCommandBuffer commandBuffer,
+                                              const VkStridedBufferRegionKHR* pRaygenShaderBindingTable,
+                                              const VkStridedBufferRegionKHR* pMissShaderBindingTable,
+                                              const VkStridedBufferRegionKHR* pHitShaderBindingTable,
+                                              const VkStridedBufferRegionKHR* pCallableShaderBindingTable, VkBuffer buffer,
+                                              VkDeviceSize offset);
+    void PostCallRecordCmdTraceRaysIndirectKHR(VkCommandBuffer commandBuffer,
+                                               const VkStridedBufferRegionKHR* pRaygenShaderBindingTable,
+                                               const VkStridedBufferRegionKHR* pMissShaderBindingTable,
+                                               const VkStridedBufferRegionKHR* pHitShaderBindingTable,
+                                               const VkStridedBufferRegionKHR* pCallableShaderBindingTable, VkBuffer buffer,
+                                               VkDeviceSize offset);
     void PostCallRecordQueueSubmit(VkQueue queue, uint32_t submitCount, const VkSubmitInfo* pSubmits, VkFence fence,
                                    VkResult result);
     void AllocateShaderPrintfResources(const VkCommandBuffer cmd_buffer, const VkPipelineBindPoint bind_point);
-    VkResult InitializeVma(VkPhysicalDevice physicalDevice, VkDevice device, VmaAllocator* pAllocator);
 };
